% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarity_data_frame.R
\name{as.data.frame.clarity_xml}
\alias{as.data.frame.clarity_xml}
\title{Return entire clarity object as tibble.}
\usage{
\method{as.data.frame}{clarity_xml}(x, ...)
}
\arguments{
\item{x}{an object of class \code{clarity_xml}}

\item{...}{arguments passed to as.data.frame}
}
\value{
a data frame, with the columns precinct, votes, votetype,
candidate, race
}
\description{
Converts the lowest geographic level (typically Precinct) of a
\code{clarity_xml} object into a long, 'tidy' dataframe.
}
