% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarity_get.R
\name{clarity_get}
\alias{clarity_get}
\title{Retrieve Clarity XML files}
\usage{
clarity_get(url)
}
\arguments{
\item{url}{The landing page URL of a clarity elections site.}
}
\value{
an object of class `clarity_xml` for manipulation/extraction
}
\description{
Retrieve Clarity XML files
}
